refactor: minor restructure and add some missing docstrings
